main_router: extract the action of updating requirement to a separate class
