+++ strictdoc/server/routers/main_router.py
@@ -1,7 +1,6 @@
 import os
 import re
 from collections import OrderedDict
-from copy import copy
 from mimetypes import guess_type
 from pathlib import Path
 from typing import Dict, List, Optional, Set, Tuple, Union
@@ -30,19 +29,11 @@ from strictdoc.backend.sdoc.models.document_grammar import (
     GrammarElement,
 )
 from strictdoc.backend.sdoc.models.object_factory import SDocObjectFactory
-from strictdoc.backend.sdoc.models.reference import (
-    ChildReqReference,
-    ParentReqReference,
-    Reference,
-)
 from strictdoc.backend.sdoc.models.requirement import (
     Requirement,
     RequirementField,
 )
 from strictdoc.backend.sdoc.models.section import Section
-from strictdoc.backend.sdoc.models.type_system import (
-    RequirementFieldName,
-)
 from strictdoc.backend.sdoc.writer import SDWriter
 from strictdoc.cli.cli_arg_parser import (
     ExportCommandConfig,
@@ -61,11 +52,12 @@ from strictdoc.core.commands.update_document_config import (
     MultipleValidationError,
     UpdateDocumentConfigCommand,
 )
+from strictdoc.core.commands.update_requirement import UpdateRequirementCommand, \
+    UpdateRequirementResult
 from strictdoc.core.document_iterator import DocumentCachingIterator
 from strictdoc.core.document_meta import DocumentMeta
 from strictdoc.core.document_tree_iterator import DocumentTreeIterator
 from strictdoc.core.project_config import ProjectConfig
-from strictdoc.core.traceability_index import RequirementConnections
 from strictdoc.export.html.document_type import DocumentType
 from strictdoc.export.html.form_objects.document_config_form_object import (
     DocumentConfigFormObject,
@@ -991,150 +983,13 @@ def create_main_router(
                 },
             )
 
-        existing_uid: Optional[str] = requirement.reserved_uid
-
-        # FIXME: [1] Leave only one method based on set_field_value().
-        # Special case: we clear out the requirement's comments and then re-fill
-        # them from scratch from the form data.
-        if "COMMENT" in requirement.ordered_fields_lookup:
-            del requirement.ordered_fields_lookup["COMMENT"]
-        if RequirementFieldName.COMMENT in requirement.ng_reserved_fields_cache:
-            del requirement.ng_reserved_fields_cache[
-                RequirementFieldName.COMMENT
-            ]
-
-        for form_field_name, form_fields in form_object.fields.items():
-            for form_field_index, form_field in enumerate(form_fields):
-                requirement.set_field_value(
-                    field_name=form_field_name,
-                    form_field_index=form_field_index,
-                    value=form_field.field_unescaped_value,
-                )
-
-        class UpdateRequirementActionObject:
-            def __init__(self):
-                self.existing_references_uids: Set[
-                    Tuple[str, Optional[str]]
-                ] = set()
-                self.reference_ids_to_remove: Set[
-                    Tuple[str, Optional[str]]
-                ] = set()
-                self.removed_uid_parent_documents_to_update: Set[
-                    Document
-                ] = set()
-                # All requirements that have to be updated. This set includes
-                # the requirement itself, all links it was linking to
-                # (for deleted links) and all links it is linking to now
-                # (including new links).
-                self.this_document_requirements_to_update: Set[
-                    Requirement
-                ] = set()
-
-        action_object = UpdateRequirementActionObject()
-        action_object.existing_references_uids.update(
-            requirement.get_parent_requirement_reference_uids()
+        update_command = UpdateRequirementCommand(
+            form_object=form_object,
+            requirement=requirement,
+            traceability_index=export_action.traceability_index,
+            config=project_config,
         )
-        action_object.reference_ids_to_remove = copy(
-            action_object.existing_references_uids
-        )
-        action_object.this_document_requirements_to_update = {requirement}
-
-        references: List[Reference] = []
-        reference_field: RequirementReferenceFormField
-        for reference_field in form_object.reference_fields:
-            ref_uid = reference_field.field_value
-            ref_type = reference_field.field_type
-            ref_role = reference_field.field_role
-            if ref_type == RequirementReferenceFormField.FieldType.PARENT:
-                references.append(
-                    ParentReqReference(
-                        parent=requirement, ref_uid=ref_uid, role=ref_role
-                    )
-                )
-            elif ref_type == RequirementReferenceFormField.FieldType.CHILD:
-                references.append(
-                    ChildReqReference(
-                        parent=requirement, ref_uid=ref_uid, role=ref_role
-                    )
-                )
-            else:
-                raise NotImplementedError(ref_type)
-        if len(references) > 0:
-            requirement.ordered_fields_lookup[RequirementFieldName.REFS] = [
-                RequirementField(
-                    parent=requirement,
-                    field_name=RequirementFieldName.REFS,
-                    field_value=None,
-                    field_value_multiline=None,
-                    field_value_references=references,
-                )
-            ]
-            requirement.references = references
-        else:
-            if RequirementFieldName.REFS in requirement.ordered_fields_lookup:
-                del requirement.ordered_fields_lookup[RequirementFieldName.REFS]
-            requirement.references = []
-
-        for (
-            document_
-        ) in export_action.traceability_index.document_tree.document_list:
-            document_.ng_needs_generation = False
-
-        # Updating Traceability Index: Links
-        for reference_field in form_object.reference_fields:
-            ref_uid = reference_field.field_value
-            ref_role: Optional[str] = (
-                reference_field.field_role
-                if reference_field.field_role is not None
-                and len(reference_field.field_role) > 0
-                else None
-            )
-            # If a link is in the form, we don't want to remove it.
-            if (ref_uid, ref_role) in action_object.reference_ids_to_remove:
-                action_object.reference_ids_to_remove.remove(
-                    (ref_uid, ref_role)
-                )
-            # If a link is already in the requirement and traceability index,
-            # there is nothing to do.
-            if (ref_uid, ref_role) in action_object.existing_references_uids:
-                continue
-            export_action.traceability_index.update_requirement_parent_uid(
-                requirement=requirement,
-                parent_uid=ref_uid,
-                role=reference_field.field_role,
-            )
-
-        # Updating Traceability Index: UID
-        export_action.traceability_index.mut_rename_uid_to_a_requirement(
-            requirement=requirement, old_uid=existing_uid
-        )
-
-        # Calculate which documents and requirements have to be regenerated.
-        for reference_id_to_remove, _ in action_object.reference_ids_to_remove:
-            removed_uid_parent_requirement = (
-                export_action.traceability_index.requirements_connections[
-                    reference_id_to_remove
-                ]
-            )
-            action_object.removed_uid_parent_documents_to_update.add(
-                removed_uid_parent_requirement.document
-            )
-            # If a link was pointing towards a parent requirement in this
-            # document, we will have to re-render it now.
-            if removed_uid_parent_requirement.document == document:
-                action_object.this_document_requirements_to_update.add(
-                    removed_uid_parent_requirement.requirement
-                )
-
-        for (
-            reference_id_to_remove,
-            reference_id_to_remove_role,
-        ) in action_object.reference_ids_to_remove:
-            export_action.traceability_index.remove_requirement_parent_uid(
-                requirement=requirement,
-                parent_uid=reference_id_to_remove,
-                role=reference_id_to_remove_role,
-            )
+        result: UpdateRequirementResult = update_command.perform()
 
         # Saving new content to .SDoc files.
         document.ng_needs_generation = True
@@ -1148,19 +1003,6 @@ def create_main_router(
         # Re-exporting HTML files.
         # Those with @ng_needs_generation == True will be regenerated.
         export_action.export()
-
-        # Rendering back the Turbo template for each changed requirement.
-        for reference_field in form_object.reference_fields:
-            ref_uid = reference_field.field_value
-            requirement_connections: RequirementConnections = (
-                export_action.traceability_index.requirements_connections[
-                    ref_uid
-                ]
-            )
-            if requirement_connections.document == document:
-                action_object.this_document_requirements_to_update.add(
-                    requirement_connections.requirement
-                )
 
         iterator: DocumentCachingIterator = (
             export_action.traceability_index.get_document_iterator(document)
@@ -1179,7 +1021,7 @@ def create_main_router(
         )
 
         output = ""
-        for requirement in action_object.this_document_requirements_to_update:
+        for requirement in result.this_document_requirements_to_update:
             template = env().get_template(
                 "actions/document/edit_requirement/"
                 "stream_update_requirement.jinja.html"
